requirements.txt
python-dotenv==1.0.0
langchain==0.1.0
langchain_openai
pydantic==2.5.2
pyodbc
azure-search-documents
azure-identity
pypdf